implemented Piazza context and integrated with DashboardPage
